changed the project_resulto result
not only chanegd teh anmes but also added class method to teh project file to create the project fiel and saved it. In fact the function has been moved from initial to the project. In this way intiial will be smaler and the function will be in its pertinent porsition.